% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordsofwisdom.R
\name{wordsofwisdom}
\alias{wordsofwisdom}
\title{Words of Wisdom}
\usage{
wordsofwisdom(wise = T)
}
\arguments{
\item{wise}{Do you need to read some words of wisdom? Defaults to TRUE.}
}
\description{
Print some words of encouragement.
}
\examples{
wordsofwisdom()
}
\keyword{encourage}
\keyword{wisdom}
