% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worklife.R
\name{worklife}
\alias{worklife}
\title{Work-Life Balance}
\usage{
worklife(after5 = T)
}
\arguments{
\item{after5}{Is it after 5:00 PM? Defaults to TRUE.}
}
\description{
Achieve the elusive work-life balance and stop working after 5 PM.
}
\examples{
worklife()
}
\keyword{balance}
\keyword{worklife}
